本来想写一个快速DCT变换的代码，发现每太搞清楚FDCT怎么做，只好放弃了，在原始的DCT变换的基础上通过先计算系数的方法加速，同时实用OMP并行加速一点点

+++ rsProcess/AerialProduct/QPDLevel1Process.cpp
@@ -1281,7 +1281,7 @@ long QPDLevel1Process::Level1Proc_ColorMatch(unsigned short* imgBuffer1, unsigne
 	}
 	GetImgHistroMatch(overlap12, overlap21, over_x, ytemp - abs(err_y), over_x, ytemp - abs(err_y), minPixel, maxPixel, mapPixel);
 #pragma omp parallel for
-	for (size_t i = 0; i < xsize1*ysize1; i++)
+	for (int i = 0; i < xsize1*ysize1; i++)
 		imgBuffer1[i] = mapPixel[imgBuffer1[i]];
 
 	////是不是考虑再做一次校正
